added author delete functionality, finalized edit author.  added list of authors to new post form

+++ server/src/controllers/authorController.ts
@@ -45,13 +45,11 @@ export const getAllAuthors = async (req: Request, res: Response) => {
 	try {
 		const authors = await pool.query("SELECT * FROM authors");
 
-		res
-			.status(200)
-			.json({
-				success: true,
-				message: "Authors fetched successfully",
-				authors: authors.rows,
-			});
+		res.status(200).json({
+			success: true,
+			message: "Authors fetched successfully",
+			authors: authors.rows,
+		});
 	} catch (error) {
 		console.error(error);
 		res.status(500).json({ error: "Internal server error" });
@@ -81,42 +79,72 @@ export const getAuthorById = async (req: Request, res: Response) => {
 // Update an author by ID
 export const updateAuthor = async (req: Request, res: Response) => {
 	const { id } = req.params;
-	const { firstName, lastName, avatarUrl, bio } = req.body;
-
-	try {
-		const updatedAuthor = await pool.query(
-			"UPDATE authors SET first_name = $1, last_name = $2, avatar_url = $3, bio = $4, updated_at = CURRENT_TIMESTAMP WHERE id = $5 RETURNING *",
-			[firstName, lastName, avatarUrl, bio, id]
-		);
-
-		if (!updatedAuthor.rows[0]) {
-			return res.status(404).json({ error: "Author not found" });
-		}
-
-		res.status(200).json(updatedAuthor.rows[0]);
-	} catch (error) {
-		console.error(error);
-		res.status(500).json({ error: "Internal server error" });
+	if (!id) {
+		return res.status(400).json({ success: false, error: "ID is required" });
 	}
+	upload.single("avatarPhoto")(req, res, async (err) => {
+		if (err) {
+			return res
+				.status(500)
+				.json({ success: false, error: "Error uploading file" });
+		}
+		const avatarUrl = (req.file as Express.MulterS3.File)?.location;
+		if (!avatarUrl) {
+			return res
+				.status(500)
+				.json({ success: false, error: "Missing avatar file" });
+		}
+		const { firstName, lastName, bio } = req.body;
+		if (!firstName || !lastName || !bio) {
+			return res
+				.status(400)
+				.json({ success: false, error: "Missing required fields" });
+		}
+		try {
+			const updatedAuthor = await pool.query(
+				"UPDATE authors SET first_name = $1, last_name = $2, avatar_url = $3, bio = $4, updated_at = CURRENT_TIMESTAMP WHERE id = $5 RETURNING *",
+				[firstName, lastName, avatarUrl, bio, id]
+			);
+
+			if (!updatedAuthor.rows[0]) {
+				return res
+					.status(404)
+					.json({ success: false, error: "Author not found" });
+			}
+
+			res.status(200).json({
+				success: true,
+				message: "Author updated successfully",
+				author: updatedAuthor.rows[0],
+			});
+		} catch (error) {
+			console.error(error);
+			res.status(500).json({ success: false, error: "Internal server error" });
+		}
+	});
 };
 
 // Delete an author by ID
 export const deleteAuthor = async (req: Request, res: Response) => {
 	const { id } = req.params;
-
+	if (!id) {
+		return res.status(400).json({ success: false, error: "ID is required" });
+	}
 	try {
 		const deletedAuthor = await pool.query(
 			"DELETE FROM authors WHERE id = $1 RETURNING *",
 			[id]
 		);
-
 		if (!deletedAuthor.rows[0]) {
-			return res.status(404).json({ error: "Author not found" });
+			return res
+				.status(404)
+				.json({ success: false, error: "Author not found" });
 		}
-
-		res.status(200).json({ message: "Author deleted successfully" });
+		res
+			.status(200)
+			.json({ success: true, message: "Author deleted successfully" });
 	} catch (error) {
 		console.error(error);
-		res.status(500).json({ error: "Internal server error" });
+		res.status(500).json({ success: false, error: "Internal server error" });
 	}
 };
